\name{Lab5Assignment-package}
\alias{Lab5Assignment-package}
\alias{Lab5Assignment}
\docType{package}
\title{
\packageTitle{Lab5Assignment}
}
\description{
\packageDescription{Lab5Assignment}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Lab5Assignment}
\packageIndices{Lab5Assignment}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{Lab5Assignment}

Maintainer: \packageMaintainer{Lab5Assignment}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the ~~
~~ R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
