% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MunicipalityResults.r
\name{MunicipalityResults}
\alias{MunicipalityResults}
\title{This function loads data from API and performs functionality based on the data.}
\usage{
MunicipalityResults(Municipality_name)
}
\arguments{
\item{Municipality_name}{Takes this as argument and party wise vote distribution is plotted.}
}
\value{
Returns the plot with party wise vote distribution of the municipality user entered.
}
\description{
In this package, we connect to a web API and fetch data from API, a functionality is created where in user inputs Municipality Name and  party wise vote distribution is plotted. The same functionality is called in shiny app making it a reactive input output program.
}
\examples{
\dontrun{ 
PartyVoteShare(Municipality_name = "Karlshamn")
}
}
