% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountyResults.R, R/MunicipalityResults.r
\docType{data}
\name{url}
\alias{url}
\title{This function loads data from API and performs functionality based on the data.}
\format{An object of class \code{character} of length 1.}
\usage{
url

url
}
\arguments{
\item{County_name}{Takes this as argument and party wise vote distribution is plotted adding all the vote shares from eah municipality in that county.}

\item{Municipality_name}{Takes this as argument and party wise vote distribution is plotted.}
}
\value{
Returns the plot with party wise vote distribution of the County user entered.

Returns the plot with party wise vote distribution of the municipality user entered.
}
\description{
In this package, we connect to a web API and fetch data from API, a functionality is created where in user inputs County Name and  party wise vote distribution is plotted adding all the vote shares from eah municipality in that count. The same functionality is called in shiny app making it a reactive input output program.

In this package, we connect to a web API and fetch data from API, a functionality is created where in user inputs Municipality Name and  party wise vote distribution is plotted. The same functionality is called in shiny app making it a reactive input output program.
}
\examples{
\dontrun{ 
CountyResults(County_name = "Dalarnas län")
}
\dontrun{ 
PartyVoteShare(Municipality_name = "Karlshamn")
}
}
\keyword{datasets}
